1.【授权下载】：增加2个按钮【新建授权】和【继续上次】，缓存上次授权下载输入的雷管； 2.【延时起爆】：延时需要减去雷管充电的时间，雷管充电的时间来自于经验统计； 3.【登录】：登录界面底下显示起爆器编号 4.【离线项目编辑】：增加缓存雷管的功能 5.【工程实施】：如果项目已经起爆成功，就不能进行编辑（包括增加雷管、连接检测、延时下载，以及编辑延时等） 5.【工程实施】：如果工程已经起爆成功，就不能真正删除； 6.【数据上报】：显示项目状态大于等于5的项目。 7. 将项目中无用的类删除

+++ app/src/main/java/com/etek/controller/activity/project/ProjectListActivity.java
@@ -82,7 +82,13 @@ public class ProjectListActivity extends BaseActivity implements View.OnClickLis
         Collections.reverse(pendingProjects);
         if (pendingProjects != null && pendingProjects.size() > 0) {
             noDataView.setVisibility(View.GONE);
-            projectInfos.addAll(pendingProjects);
+            for(PendingProject pp:pendingProjects){
+                Log.d(TAG,"项目："+pp.getId() +"\t项目状态："+pp.getProjectStatus()+"\t上报状态："+pp.getReportStatus());
+                if(pp.getProjectStatus()<AppIntentString.PROJECT_IMPLEMENT_DATA_DELETE){
+                    projectInfos.add(pp);
+                }
+            }
+            //projectInfos.addAll(pendingProjects);
         } else {
             noDataView.setVisibility(View.VISIBLE);
         }
@@ -237,10 +243,29 @@ public class ProjectListActivity extends BaseActivity implements View.OnClickLis
         builder.setPositiveButton("确定", new DialogInterface.OnClickListener() {
             @Override
             public void onClick(DialogInterface dialog, int which) {
-                //删除数据
-                DBManager.getInstance().getPendingProjectDao().delete(pendingProject);
-                List<ProjectDetonator> projectDetonators = DBManager.getInstance().getProjectDetonatorDao()._queryPendingProject_DetonatorList(pendingProject.getId());
-                DBManager.getInstance().getProjectDetonatorDao().deleteInTx(projectDetonators);
+
+                Log.d(TAG,"项目："+pendingProject.getId()
+                        + "项目状态："+ pendingProject.getProjectStatus()
+                        + "上报状态："+ pendingProject.getReportStatus());
+
+                if(pendingProject.getProjectStatus()<AppIntentString.PROJECT_IMPLEMENT_DATA_REPORT1){
+                    // 如果还未起爆，就可以删除项目
+                    DBManager.getInstance().getPendingProjectDao().delete(pendingProject);
+
+                    List<ProjectDetonator> projectDetonators = DBManager.getInstance().getProjectDetonatorDao()._queryPendingProject_DetonatorList(pendingProject.getId());
+                    DBManager.getInstance().getProjectDetonatorDao().deleteInTx(projectDetonators);
+                }else{
+                    if((pendingProject.getProjectStatus()==AppIntentString.PROJECT_IMPLEMENT_DATA_REPORT1)
+                        &&(!pendingProject.getReportStatus().equals("1"))){
+                        // 已经起爆了，但未上报成功
+                        showDialogMessage("项目未上报成功，不能删除！");
+                        return;
+                    }else{
+                        // 更新项目状态为  PROJECT_IMPLEMENT_DATA_DELETE
+                        pendingProject.setProjectStatus(AppIntentString.PROJECT_IMPLEMENT_DATA_DELETE);
+                        DBManager.getInstance().getPendingProjectDao().save(pendingProject);
+                    }
+                }
                 projectInfos.remove(position);
                 projectListAdapter.notifyDataSetChanged();
                 dialog.dismiss();
